[ax] bugfix: prevent hang on slow callback response
if the rx callback takes a long time to return, the FIFO may
completely fill such that the register FIFOCOUNT = 256 when
ax_fifo_rx_data is eventually called again. This value was locally
stored in a uint8_t, and thus the value 256 was stored as 0. Therefore
ax_fifo_rx_data made no further attempts to read the FIFO, and no data
was received until the radio was reset. Fix with correct local storage
type of uint16_t.